MQTT: add auth and topic filtering
Petnet devices will get a unique API key with which they will log into
the MQTT broker. The Petnet devices will also only be able to subscribe
to the topics of their own gateway IDs.

The server has its own login that is generated whenever the server
starts up. That login can subscribe to any topic in order for it to
respond to all feeders.

Currently there appears to be no way to filter what topics can be
published by a client.